+++ packages/cli/src/store/index.ts
@@ -16,7 +16,7 @@ import type {CliActionType, CliOptionsType, EMPModeType} from 'src/types/env'
 import {chainName} from './chain'
 import {HtmlEmpInjectPlugin} from './rspack/builtInPlugin'
 export class GlobalStore {
-  public rspack = rspack
+  public rspack: typeof rspack = rspack
   public rspackVersion = rspack.rspackVersion
   public isOldRspack = vCompare(this.rspackVersion, '1.0.0') === -1
   /**